+++ shared/src/business/entities/trialSessions/SpecialTrialSessions.ts
